Modularize CLI architecture and enforce robust baseline management
- Extract GitHub API logic into `dev-tools/gh_client.py`.
- Migrate issue validation to `dev-tools/issue_manager.py`.
- Migrate baseline/ratchet logic to `dev-tools/baseline_manager.py`.
- Migrate PR/Review lifecycle logic to `dev-tools/pr_manager.py`.
- Enforce raw JSON output in `scripts/detect-antipatterns.mjs` when `--json` is used.
- Update `tests/dev-tools/test_td_cli.py` to match new module structure.
- Refine documentation in `CONTRIBUTING.md` and `AGENTS.md`.
