+++ dev-tools/issue_manager.py
@@ -1,0 +1,105 @@
+import re
+import json
+import sys
+from utils import get_repo_name, CLIError
+from gh_client import get_github_token
+
+EXISTING_COMPONENTS = {
+    'Box': 'src/layouts/Box.tsx', 'Stack': 'src/layouts/Stack.tsx', 'Grid': 'src/layouts/Grid.tsx',
+    'Text': 'src/layouts/Text.tsx', 'Button': 'src/layouts/Button.tsx', 'ContentCard': 'src/components/ui/ContentCard.tsx',
+    'PageHeader': 'src/components/ui/PageHeader.tsx', 'FilterBar': 'src/components/ui/FilterBar.tsx',
+    'FolioGrid': 'src/components/ui/FolioGrid.tsx', 'Skeleton': 'src/components/ui/Skeleton.tsx',
+    'ViewToggle': 'src/components/ui/ViewToggle.tsx', 'ListRow': 'src/components/ui/ListRow.tsx',
+    'MarkdownRenderer': 'src/components/ui/MarkdownRenderer.tsx', 'DetailLayout': 'src/components/layout/DetailLayout.tsx',
+    'useSearchParam': 'src/hooks/useSearchParam.ts', 'useHotkeys': 'src/hooks/useHotkeys.ts', 'safeSearch': 'src/lib/utils.ts',
+}
+
+BANNED_PATTERNS = [
+    (r'HashRouter', 'HashRouter is banned. Use createBrowserRouter (AGENTS.md §9)'),
+    (r'import React from .react.', 'Unnecessary React import — React 17+ (AGENTS.md §4)'),
+    (r'style=\{\{', 'Inline styles are banned. Use design tokens (AGENTS.md §11)'),
+    (r'text-\[\d+px\]', 'Arbitrary px Tailwind value. Use design tokens (AGENTS.md §1)'),
+    (r'bg-\[#', 'Raw hex color in Tailwind. Use CSS variables from tokens.css'),
+    (r'<div\s+className=".*?(flex|grid|p-|m-)', 'Raw layout div. Use <Box/>, <Stack/>, <Grid/> primitives (AGENTS.md §3)'),
+    (r'className=".*?text-\[\d', 'Arbitrary text size. Use typeSizes from design-tokens.ts'),
+]
+
+RENAMED_ASSETS = { 'accent-brand': 'accent', 'useSearch': 'useSearchParam' }
+DEPRECATED_PATHS = { 'src/components/common/': 'src/components/ui/' }
+REQUIRED_FOR_CONTENT_ISSUES = ['type', 'title', 'date', 'author', 'category', 'excerpt']
+
+def extract_code_blocks(text: str) -> list[str]:
+    return re.findall(r'```(?:tsx?|jsx?|html)?\n(.*?)```', text, re.DOTALL)
+
+def handle_validate_issue(args):
+    from github import Github
+    token = get_github_token()
+    if not token: raise CLIError("GitHub token not found", code=401)
+    repo = Github(token).get_repo(get_repo_name())
+
+    issues = []
+    if args.all_open: issues = list(repo.get_issues(state='open'))
+    elif args.issue_number: issues = [repo.get_issue(args.issue_number)]
+    else: raise CLIError("Provide --issue-number or --all-open")
+
+    results = []
+    total_findings = 0
+    for issue in issues:
+        findings, warnings = [], []; body = issue.body or ''; title = issue.title or ''
+        for i, block in enumerate(extract_code_blocks(body)):
+            for pattern, msg in BANNED_PATTERNS:
+                if re.search(pattern, block): findings.append(f"Code block {i+1}: {msg}")
+            for comp, path in EXISTING_COMPONENTS.items():
+                if re.search(rf'(create|build|make|add|new)\s+.*{comp}', block, re.IGNORECASE): warnings.append(f"Code block {i+1}: Suggests `{comp}` (exists at `{path}`)")
+        for comp, path in EXISTING_COMPONENTS.items():
+            if re.search(rf'(create|build|make|add\s+a\s+new)\s+.*{comp}\b', body, re.IGNORECASE): warnings.append(f"Issue suggests `{comp}` (exists at `{path}`)")
+        if title.startswith('Draft:') and '```markdown' in body:
+            md_match = re.search(r'```markdown\n(.*?)\n```', body, re.DOTALL)
+            if md_match:
+                for field in REQUIRED_FOR_CONTENT_ISSUES:
+                    if not re.search(rf'^{field}:', md_match.group(1), re.MULTILINE): findings.append(f"Missing frontmatter: `{field}`")
+        if not re.search(r'(acceptance criteria|definition of done|## done|verify|test)', body, re.IGNORECASE): warnings.append("No acceptance criteria.")
+        if re.search(r'tailwind|className.*flex|className.*grid', body, re.IGNORECASE) and not re.search(r'<Box|<Stack|<Grid|primitives|design.tokens', body, re.IGNORECASE): warnings.append("Mentions Tailwind but not layout primitives.")
+
+        issue_result = {"number": issue.number, "title": title, "findings": findings, "warnings": warnings}
+        results.append(issue_result)
+        total_findings += len(findings)
+
+        if not args.json:
+            print(f"{'✅' if not findings else '❌'} #{issue.number}: {title[:60]}")
+            for f in findings: print(f"   ❌ {f}")
+            for w in warnings: print(f"   ⚠️  {w}")
+
+        if args.post_comments and (findings or warnings):
+            comment = "## 🤖 Issue Quality Review\n\n"
+            if findings: comment += "### ❌ Violations\n" + "\n".join(f"- {f}" for f in findings) + "\n\n"
+            if warnings: comment += "### ⚠️ Warnings\n" + "\n".join(f"- {w}" for w in warnings) + "\n"
+            if not args.dry_run: issue.create_comment(comment + "\n---\n*Generated by `td_cli validate-issue`*")
+
+    if args.json: print(json.dumps({"status": "success" if total_findings == 0 else "error", "issues": results}, indent=2))
+    if total_findings > 0: sys.exit(1)
+
+def handle_update_issues(args):
+    from github import Github
+    token = get_github_token(); repo_name = get_repo_name()
+    if not token: raise CLIError("GitHub token not found", code=401)
+    g = Github(token); repo = g.get_repo(repo_name)
+
+    updates = []
+    if not args.json: print(f"🔍 Scanning open issues in {repo_name}...")
+    for issue in repo.get_issues(state='open'):
+        body = issue.body or ''; findings = []
+        for old, new in RENAMED_ASSETS.items():
+            if old in body: findings.append(f"References deprecated name `{old}`. Use `{new}` instead.")
+        for old, new in DEPRECATED_PATHS.items():
+            if old in body: findings.append(f"References deprecated path `{old}`. New location: `{new}`")
+        for pattern, message in BANNED_PATTERNS:
+            if re.search(pattern, body): findings.append(f"Contains banned pattern: {message}")
+        if findings:
+            updates.append({"number": issue.number, "findings": findings})
+            comment = "## 🤖 Automated Issue Update\n\n" + "\n".join(f"- {f}" for f in findings) + "\n\n---\n*Generated by `td_cli update-issues`*"
+            if not args.json: print(f"[{'DRY-RUN' if args.dry_run else 'EXECUTE'}] Found {len(findings)} issues in #{issue.number}")
+            if not args.dry_run: issue.create_comment(comment); print(f"✅ Posted update comment to #{issue.number}")
+            elif not args.json: print(f"Preview for #{issue.number}:\n{comment}\n")
+
+    if args.json: print(json.dumps({"status": "success", "updates": updates}, indent=2))
