+++ dev-tools/pr_manager.py
@@ -1,0 +1,167 @@
+import os
+import re
+import json
+import subprocess
+from collections import defaultdict
+from utils import get_repo_name, CLIError
+from gh_client import get_github_token
+from scope_check import verify_pr_scope
+
+def get_pr_files(pr) -> set[str]:
+    return {f.filename for f in pr.get_files()}
+
+def detect_conflicts(repo, target_pr_num=None):
+    open_prs = list(repo.get_pulls(state='open'))
+    file_to_prs = defaultdict(list)
+    for pr in open_prs:
+        for f in get_pr_files(pr): file_to_prs[f].append(pr.number)
+
+    conflicts = defaultdict(list)
+    for filename, prs in file_to_prs.items():
+        if len(prs) > 1 and (target_pr_num is None or target_pr_num in prs):
+            conflicts[tuple(sorted(prs))].append(filename)
+    return conflicts
+
+def handle_conflicts(args):
+    from github import Github
+    token = get_github_token()
+    if not token: raise CLIError("GitHub token not found", code=401)
+    repo = Github(token).get_repo(get_repo_name())
+    conflicts = detect_conflicts(repo, args.pr)
+
+    formatted = []
+    for pr_pair, files in conflicts.items():
+        formatted.append({"prs": list(pr_pair), "files": files})
+        if not args.json:
+            print(f"⚠️  {' ↔ '.join(f'#{p}' for p in pr_pair)} share {len(files)} file(s):")
+            for f in sorted(files)[:10]: print(f"    - {f}")
+
+    if args.json: print(json.dumps({"status": "success", "conflicts": formatted}, indent=2))
+    elif not conflicts: print("✅ No potential merge conflicts detected.")
+
+def handle_status_board(args):
+    from github import Github
+    token = get_github_token()
+    if not token: raise CLIError("GitHub token not found", code=401)
+    repo = Github(token).get_repo(get_repo_name())
+
+    prs_data = []
+    if not args.json: print("# Active Agent Work Board\n| Branch | Issue | Status | Conflicts |\n|--------|-------|--------|-----------|")
+
+    for pr in repo.get_pulls(state='open'):
+        m = re.search(r'issue-(\d+)', pr.head.ref); issue = f"#{m.group(1)}" if m else "—"
+        if not args.json: print(f"| {pr.head.ref} | {issue} | {'Draft' if pr.draft else 'Open'} | ... |")
+        prs_data.append({"branch": pr.head.ref, "issue": issue, "status": "Draft" if pr.draft else "Open"})
+
+    if args.json: print(json.dumps({"status": "success", "work": prs_data}, indent=2))
+
+def handle_audit_pr(args):
+    pr_num = args.pr_number; review_dir = os.path.join(os.getcwd(), "dev-tools", "logs", "reviews")
+    ctx_path = os.path.join(review_dir, f"pr-context-{pr_num}.md"); rev_path = os.path.join(review_dir, f"pr-review-{pr_num}.md")
+
+    res = {"pr": pr_num, "files": {}}
+
+    if args.fetch:
+        token = get_github_token()
+        if not token: raise CLIError("GitHub token not found", code=401)
+        from github import Github; repo = Github(token).get_repo(get_repo_name()); pr = repo.get_pull(int(pr_num))
+        title = pr.title; author = pr.user.login; desc = pr.body or '_No description provided._'
+        context_lines = [f"# PR Context: #{pr.number} — {title}", f"**Author:** @{author}\n", f"## Description\n{desc}\n", "## Files Changed"]
+        for f in pr.get_files(): context_lines.append(f"- {'🟢' if f.status=='added' else '🔴' if f.status=='removed' else '🟡'} `{f.filename}`")
+        context_lines.append("\n## Diffs")
+        for f in pr.get_files():
+            context_lines.append(f"\n### `{f.filename}` ({f.status})")
+            patch = f.patch or '_No textual diff available._'; annotated = []; line_num = 0
+            if patch != '_No textual diff available._':
+                for line in patch.splitlines():
+                    if line.startswith('@@'):
+                        m = re.search(r'\+(\d+)', line); line_num = int(m.group(1)) if m else line_num
+                        annotated.append(line)
+                    elif line.startswith('+'): annotated.append(f"{line_num:4d} |{line}"); line_num += 1
+                    elif line.startswith('-'): annotated.append(f"     |{line}")
+                    else: annotated.append(f"{line_num:4d} |{line}"); line_num += 1
+            context_lines.append(f"```diff\n" + "\n".join(annotated) + "\n```")
+        os.makedirs(review_dir, exist_ok=True)
+        with open(ctx_path, "w") as f: f.write("\n".join(context_lines))
+        template_path = os.path.join(os.path.dirname(__file__), "review_template.md")
+        if os.path.exists(template_path): template = open(template_path).read().format(pr_num=pr_num, head_sha=pr.head.sha)
+        else: template = f"# PR Review: #{pr_num}\n- SHA: {pr.head.sha}\n"
+        with open(rev_path, "w") as f: f.write(template)
+        res["files"]["context"] = ctx_path
+        res["files"]["review"] = rev_path
+        if not args.json: print(f"✅ Generated review files for PR #{pr_num}")
+
+    if args.audit:
+        if not os.path.exists(ctx_path): raise CLIError(f"Context file missing: {ctx_path}")
+        with open(ctx_path) as f: context = f.read()
+        changed_files = re.findall(r'### `([^`]+)`', context); auto_findings = []
+
+        scope_warning = verify_pr_scope(changed_files)
+        if scope_warning:
+            auto_findings.append({"path": "PR SCOPE", "issue": scope_warning, "severity": "major"})
+
+        files_to_audit = [f for f in changed_files if (f.endswith('.tsx') or f.endswith('.ts')) and os.path.exists(f)]
+        if files_to_audit:
+            try:
+                # Use pnpm run audit as requested, passing targets after --
+                proc = subprocess.run(["pnpm", "run", "audit", "--", "--json"] + files_to_audit, capture_output=True, text=True)
+                if proc.stdout:
+                    from utils import extract_json
+                    audit_data = extract_json(proc.stdout)
+                    if audit_data:
+                        for filepath, violations in audit_data.items():
+                            for v in violations:
+                                auto_findings.append({
+                                    "path": filepath,
+                                    "issue": f"{v['pattern']}: {v['message']} (value: {v['value']})",
+                                    "severity": v.get('severity', 'minor')
+                                })
+            except Exception as e:
+                if not args.json: print(f"⚠️  Audit script failed: {e}")
+
+        res["auto_findings"] = auto_findings
+        if not args.json:
+            if auto_findings:
+                print(f"📋 Found {len(auto_findings)} violations:")
+                for f in auto_findings: print(f"  [{f['severity'].upper()}] {f['path']}: {f['issue']}")
+            subprocess.call(["copilot", "-p", f"Auditing PR #{pr_num}...", "--allow-tool", "read", "--allow-tool", "write", "--allow-tool", "file_edit"])
+
+    if args.submit:
+        from submit_review import submit_review
+        submit_review(pr_num, rev_path, cleanup=args.cleanup, dry_run=args.dry_run, event_override=args.event, is_json=args.json)
+
+    if args.json: print(json.dumps({"status": "success", "data": res}, indent=2))
+
+def handle_manage_reviews(args):
+    from github import Github
+    token = get_github_token()
+    if not token: raise CLIError("GitHub token not found", code=401)
+    g = Github(token); repo = g.get_repo(get_repo_name()); login = g.get_user().login
+
+    prs_data = []
+    for pr in repo.get_pulls(state='open', sort='updated', direction='desc'):
+        last_review = next((r for r in pr.get_reviews().reversed if r.user.login == login), None)
+        status = "ACTION: Needs Review" if not last_review else f"ACTION: Needs Re-Review" if last_review.commit_id != pr.head.sha else "STATE: Up-To-Date"
+
+        pr_item = {"number": pr.number, "title": pr.title, "status": status, "unaddressed": []}
+        if not args.json: print(f"[PR #{pr.number}] {pr.title}\n  ├── {status}")
+
+        if args.check_responses:
+            revs = list(pr.get_reviews()); our_revs = [r for r in revs if r.user.login == login]
+            if our_revs:
+                last_time = max(r.submitted_at for r in our_revs); our_coms = [c for c in pr.get_review_comments() if c.user.login == login]
+                after_coms = [c for c in pr.get_review_comments() if c.user.login != login and any(c.in_reply_to_id == oc.id for oc in our_coms)]
+                commits_after = [c for c in pr.get_commits() if c.commit.author.date > last_time]
+                for oc in our_coms:
+                    if not any(ac.in_reply_to_id == oc.id for ac in after_coms) and not commits_after:
+                        pr_item["unaddressed"].append(f"{oc.path}:{oc.position}")
+                if pr_item["unaddressed"] and not args.json: print(f"  └── ⚠️ UNADDRESSED ({len(pr_item['unaddressed'])})")
+
+        if args.cleanup_comments:
+            for c in pr.get_issue_comments():
+                if c.user.login == login and "<!-- td-review-manager-comment -->" in c.body:
+                    if not args.dry_run: c.delete(); print(f"  Deleted tool comment {c.id}")
+                    elif not args.json: print(f"  [DRY-RUN] Would delete tool comment {c.id}")
+        prs_data.append(pr_item)
+
+    if args.json: print(json.dumps({"status": "success", "prs": prs_data}, indent=2))
