+++ tests/dev-tools/test_td_cli.py
@@ -12,8 +12,8 @@ from submit_review import submit_review
 
 class TestTDCLI(unittest.TestCase):
 
-    @patch('td_cli.get_github_token')
-    @patch('td_cli.get_repo_name')
+    @patch('issue_manager.get_github_token')
+    @patch('issue_manager.get_repo_name')
     @patch('github.Github')
     def test_validate_issue_dry_run_default(self, mock_github_class, mock_repo, mock_token):
         """Test that validate-issue defaults to dry-run True"""
